requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# AI & NLP
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.35.0
torch>=2.2.0
sentence-transformers>=2.2.0
bertopic>=0.15.0

# Data Processing
polars>=0.20.0
duckdb>=0.9.0
pandas>=2.1.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Storage
minio>=7.2.0
boto3>=1.34.0

# Orchestration
prefect>=2.14.0
celery>=5.3.0
schedule>=1.2.0

# Monitoring
prometheus-client>=0.19.0

# Web Scraping
beautifulsoup4>=4.12.0
httpx>=0.25.0

# Time Series & Prediction
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Visualization
wordcloud>=1.9.0
matplotlib>=3.7.0
Pillow>=10.0.0
plotly>=5.15.0
